# bigdata-vis/api/swagger/swagger (copia).yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Big Data Visualization API
# during dev, should point to your local machine
host: localhost:20183
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /datavis/scatterplot:
    x-swagger-router-controller: scatterplot
    get:
      description: Returns information about the requested scatterplot
      operationId: scatterplot
      parameters:
        - name: file
          in: query
          description: Path to file
          required: true
          type: string
        - name: colx
          in: query
          description: The column selected for the scatterplot
          required: true
          type: string
        - name: coly
          in: query
          description: The column selected for the scatterplot
          required: true
          type: string
        - name: secx
          in: query
          description: Number of sections to divide values for the x-axis
          required: true
          type: number
        - name: secy
          in: query
          description: Number of sections to divide values for the y-axis
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ScatterplotResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /datavis/linechart:
    x-swagger-router-controller: linechart
    get:
      description: Returns information about the requested linechart
      operationId: linechart
      parameters:
        - name: file
          in: query
          description: Path to file
          required: true
          type: string
        - name: sec
          in: query
          description: Number of sections to divide values for the axis
          required: true
          type: number
        - name: colX
          in: query
          description: The column selected for the X-Axis
          required: true
          type: string
        - name: count
          in: query
          description: Number of column selected
          required: true
          type: number
        - name: col
          in: query
          description: Array with column name selected
          required: true
          type: array
          items:
            type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LinechartResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /datavis/stackedAreaChart:
    x-swagger-router-controller: stackedAreaChart
    get:
      description: Returns information about the requested stackedAreaChart
      operationId: stackedAreaChart
      parameters:
        - name: file
          in: query
          description: Path to file
          required: true
          type: string
        - name: sec
          in: query
          description: Number of sections to divide values for the axis
          required: true
          type: number
        - name: colX
          in: query
          description: The column selected for the X-Axis
          required: true
          type: string
        - name: count
          in: query
          description: Number of column selected
          required: true
          type: number
        - name: col
          in: query
          description: Array with column name selected
          required: true
          type: array
          items:
            type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/StackedAreaChartResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ScatterplotResponse:
    properties:
      _id:
        type: string
      title:
        type: string
      titleX:
        type: string
      titleY:
        type: string
      values:
        type: array
        items:
          type: object
          properties:
            pointX:
              type: number
            pointY:
              type: number
            value:
              type: number
            valueColor:
              type: number
...
  LinechartResponse:
    properties:
      _id:
        type: string
      title:
        type: string
      colX:
        type: string
      listLines:
        type: array
        items:
          type: object
          properties:
            titleLine:
              type: string
            lvalues:
              type: array
              items:
                type: object
                properties:
                  pointXMin:
                    type: number
                  pointXMax:
                    type: number
                  pointY:
                    type: number
  StackedAreaChartResponse:
    properties:
      _id:
        type: string
      title:
        type: string
      colX:
        type: string
      listLines:
        type: array
        items:
          type: object
          properties:
            titleLine:
              type: string
            lvalues:
              type: array
              items:
                type: object
                properties:
                  pointXMin:
                    type: number
                  pointXMax:
                    type: number
                  pointY:
                    type: number
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string